requirements.txt
asgiref==3.4.1; python_version >= '3.6'
django==3.2.9
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
psycopg2-binary==2.9.2
pyjwt==2.3.0; python_version >= '3.6'
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
